+++ src/main/java/nbc/mushroom/domain/like/service/AuctionItemLikeService.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import nbc.mushroom.domain.auction_item.entity.AuctionItem;
 import nbc.mushroom.domain.auction_item.repository.AuctionItemRepository;
 import nbc.mushroom.domain.common.exception.CustomException;
+import nbc.mushroom.domain.like.dto.response.CheckLikedAuctionItemRes;
 import nbc.mushroom.domain.like.entity.AuctionItemLike;
 import nbc.mushroom.domain.like.repository.AuctionItemLikeRepository;
 import nbc.mushroom.domain.user.entity.User;
@@ -45,6 +46,15 @@ public class AuctionItemLikeService {
             .build();
 
         auctionItemLikeRepository.save(auctionItemLike);
+    }
+
+    // 본인이 해당 경매 물품에, 좋아요를 한 여부를 확인
+    public CheckLikedAuctionItemRes getLikedAuctionItem(User user, Long auctionItemId) {
+
+        Boolean hasLike = auctionItemLikeRepository.existAuctionItemLikeByUserAndAuctionItem(
+            user, auctionItemId);
+
+        return new CheckLikedAuctionItemRes(hasLike);
     }
 
     public void hardDeleteAuctionItemLike(Long userId, Long auctionItemId) {
