+++ src/main/java/nbc/mushroom/domain/like/service/LikedAuctionItemService.java
@@ -2,7 +2,6 @@ package nbc.mushroom.domain.like.service;
 
 
 import lombok.RequiredArgsConstructor;
-import nbc.mushroom.domain.like.dto.response.CheckLikedAuctionItemRes;
 import nbc.mushroom.domain.like.dto.response.LikedAuctionItemRes;
 import nbc.mushroom.domain.like.repository.AuctionItemLikeRepository;
 import nbc.mushroom.domain.user.entity.User;
@@ -22,15 +21,6 @@ public class LikedAuctionItemService {
 
         return auctionItemLikeRepository.findAuctionItemLikeByUserId(
             user, pageable);
-    }
-
-    // 본인이 해당 경매 물품에, 좋아요를 한 여부를 확인
-    public CheckLikedAuctionItemRes getLikedAuctionItem(User user, Long auctionItemId) {
-
-        Boolean hasLike = auctionItemLikeRepository.existAuctionItemLikeByUserAndAuctionItem(
-            user, auctionItemId);
-
-        return new CheckLikedAuctionItemRes(!hasLike);
     }
 
 }
